Airport loader and class and test fixes.

+++ src/main/java/project/model/Airport.java
@@ -11,14 +11,14 @@ public class Airport {
     private double latitude;
     private double longitude;
     private int altitude;
-    private int timezone;
+    private double timezone;
     private String dst;
     private String timezoneString;
     private String type;
     private String source;
     private int numRoutes;
 
-    public Airport(int id, int risk, String name, String city, String country, String iata, String icao, double latitude, double longitude, int altitude, int timezone, String dst, String timezoneString, String type, String source, int numRoutes) {
+    public Airport(int id, int risk, String name, String city, String country, String iata, String icao, double latitude, double longitude, int altitude, double timezone, String dst, String timezoneString, String type, String source, int numRoutes) {
         setId(id);
         setRisk(risk);
         setName(name);
@@ -117,11 +117,11 @@ public class Airport {
         this.altitude = altitude;
     }
 
-    public int getTimezone() {
+    public double getTimezone() {
         return timezone;
     }
 
-    public void setTimezone(int timezone) {
+    public void setTimezone(double timezone) {
         this.timezone = timezone;
     }
